% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_cv.R
\name{calc_cv}
\alias{calc_cv}
\title{Computes the coefficient of variation of a numeric vector}
\usage{
calc_cv(x)
}
\arguments{
\item{x}{input numeric vector}
}
\value{
A numeric value of the relative standard deviation
}
\description{
Computes the coefficient of variation of a numeric vector
}
\examples{
calc_cv(1:10)
}
