% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cv.R
\name{get_cv}
\alias{get_cv}
\title{Perform a row-wise relative standard deviation calculation and return a numeric vector of the coefficient of variation}
\usage{
get_cv(mat)
}
\arguments{
\item{mat}{input data matrix}
}
\value{
A numeric vector of the coefficient of variation
}
\description{
Perform a row-wise relative standard deviation calculation and return a numeric vector of the coefficient of variation
}
\examples{
mat <- matrix(rnorm(50),5,10)
get_cv(mat)
}
