% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pgroup.R
\name{get_pgroup}
\alias{get_pgroup}
\title{Peptide/Protein group inference}
\usage{
get_pgroup(dat, protGrp = F, average = T)
}
\arguments{
\item{dat}{input dataframe}

\item{protGrp}{logical variable indicating if protein groups should be inferred}

\item{average}{logical variable indicating if abundances should be averaged}
}
\value{
dataframe with the peptide/protein groups inferred and abundances averaged or summed
}
\description{
Peptide/Protein group inference
}
\examples{
pepnames <- c("a@A", "b@A", "b@B", "c@C", "d@C", "a@D", "b@D")
dat <- matrix(runif(35, min=8, max=25), 7, 5)
rownames(dat) <- pepnames
pepGrp <- get_pgroup(dat)
}
