% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_supfig1_2.r
\name{get_supfig1_2}
\alias{get_supfig1_2}
\title{Sci Data Paper Supplementary Figures 1 and 2}
\usage{
get_supfig1_2()
}
\value{
supplFigure1_2.pdf
}
\description{
Sci Data Paper Supplementary Figures 1 and 2
}
