% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_median.R
\name{normalize_median}
\alias{normalize_median}
\title{Median location Normalization}
\usage{
normalize_median(df)
}
\arguments{
\item{df}{Numeric matrix or dataframe}
}
\value{
Global median location normalized numeric matrix
}
\description{
Median location Normalization
}
\examples{
mat <- matrix(rnorm(50),5,10)
normalize_median(mat)
}
