% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_to_raw.R
\name{source_to_raw}
\alias{source_to_raw}
\title{Process source data to raw}
\usage{
source_to_raw(dat)
}
\arguments{
\item{dat}{source peak area output from Skyline}
}
\value{
formatted raw dataframe
}
\description{
Process source data to raw
}
